requirements.txt
flask==3.0.0
google-generativeai==0.3.2
chromadb==0.4.22
python-dotenv==1.0.0
sentence-transformers==2.2.2